+++ helpers/NPVersions.js
@@ -39,6 +39,7 @@ export function usersVersionHas(feature: string): boolean {
     reuseSplitView: '3.20.1', // Jan 2026, macOS build 1479ish
     windowIsVisible: '3.20.2', // Mar 2026, macOS build 1494
     commandBarForms: '3.21.0', // Apr 2026, macOS build 1502
+    APIsAvailableInWebViews: '3.21.0', // ~ Apr 2026
   }
 
   // Check if the user's version meets the requirement for the requested feature